+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added support for Python 3.13.
 * Dropped support for Python 3.8 which has reached end of life. Python 3.9+ is
   now required.
+* `PipelineApiClient.get_pipeline_instances` now supports querying pipeline
+  instances by dataset ID and version.
 
 ## 3.2.0 - 2024-10-03